feat: architecture improvements — centralized query config, toast system, error boundary, category parser refactor
- Extract query cache timing constants into shared queryConfig.ts (eliminates magic numbers across 10 files)
- Add toast notification system (Zustand store + Toast component) with mutation feedback
- Add ErrorBoundary wrapping route outlet for graceful error recovery
- Split 877-line categoryParser.ts into modular category/ directory (types, helpers, mappings, detector, grouper)
- Extract useContentRailData hook to deduplicate rail merge/sort/limit logic from home/api.ts
- Add FocusableElement generic wrapper, simplify FocusableCard
- Performance: memo() on ContentRail/HorizontalScroll, DOM data-attribute for inputMode (avoids 110+ re-renders)
- Remove suppressArrowNav/isPlaying store fields, derive from player state directly
- Add 41 unit tests (store, useDebounce, formatDuration, isNewContent)
