+++ src/features/player/components/PlayerPage.tsx
@@ -6,7 +6,7 @@ import { PlayerOSD, type OSDAction } from './PlayerOSD';
 import { usePlayerKeyboard } from '../hooks/usePlayerKeyboard';
 import { isTVMode } from '@shared/utils/isTVMode';
 import { useProgressTracking } from '../hooks/useProgressTracking';
-import { usePlayerStore, useUIStore } from '@lib/store';
+import { usePlayerStore } from '@lib/store';
 import {
   useSpatialContainer,
   FocusContext,
@@ -137,22 +137,20 @@ export function PlayerPage({
   });
 
   // In TV mode: pause spatial nav (player handles seek/volume via usePlayerKeyboard),
-  // and auto-focus play/pause when player mounts
-  const setSuppressArrowNav = useUIStore((s) => s.setSuppressArrowNav);
+  // and auto-focus play/pause when player mounts.
+  // Arrow suppression is derived from player active state + isTVMode in SpatialNavProvider.
   useEffect(() => {
     if (isTVMode) {
-      setSuppressArrowNav(true);
       pauseSpatialNav();
       const timer = setTimeout(() => {
         try { setFocus('player-play-pause'); } catch { /* not registered yet */ }
       }, 100);
       return () => {
         clearTimeout(timer);
-        setSuppressArrowNav(false);
         resumeSpatialNav();
       };
     }
-  }, [streamId, setSuppressArrowNav]);
+  }, [streamId]);
 
   // Keep controls visible if user is navigating controls with D-pad
   useEffect(() => {
