+++ src/features/vod/api.ts
@@ -1,12 +1,13 @@
 import { useQuery } from '@tanstack/react-query';
 import { api } from '@lib/api';
+import { STALE_TIMES } from '@lib/queryConfig';
 import type { XtreamCategory, XtreamVODStream, XtreamVODInfo } from '@shared/types/api';
 
 export function useVODCategories() {
   return useQuery({
     queryKey: ['vod', 'categories'],
     queryFn: () => api<XtreamCategory[]>('/vod/categories'),
-    staleTime: 6 * 60 * 60 * 1000,
+    staleTime: STALE_TIMES.categories,
   });
 }
 
@@ -15,7 +16,7 @@ export function useVODStreams(categoryId: string) {
     queryKey: ['vod', 'streams', categoryId],
     queryFn: () => api<XtreamVODStream[]>(`/vod/streams/${categoryId}`),
     enabled: !!categoryId,
-    staleTime: 2 * 60 * 60 * 1000,
+    staleTime: STALE_TIMES.streams,
   });
 }
 
@@ -24,6 +25,6 @@ export function useVODInfo(vodId: string) {
     queryKey: ['vod', 'info', vodId],
     queryFn: () => api<XtreamVODInfo>(`/vod/info/${vodId}`),
     enabled: !!vodId,
-    staleTime: 2 * 60 * 60 * 1000,
+    staleTime: STALE_TIMES.streams,
   });
 }
