+++ src/lib/queryClient.ts
@@ -1,10 +1,11 @@
 import { QueryClient } from '@tanstack/react-query';
+import { STALE_TIMES, GC_TIMES } from './queryConfig';
 
 export const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
-      staleTime: 5 * 60 * 1000, // 5 minutes
-      gcTime: 30 * 60 * 1000,   // 30 minutes
+      staleTime: STALE_TIMES.default,
+      gcTime: GC_TIMES.default,
       retry: 1,
       refetchOnWindowFocus: false,
     },
