+++ src/shared/hooks/__tests__/useDebounce.test.ts
@@ -1,0 +1,80 @@
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { renderHook, act } from '@testing-library/react';
+import { useDebounce } from '../useDebounce';
+
+describe('useDebounce', () => {
+  beforeEach(() => {
+    vi.useFakeTimers();
+  });
+
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it('returns initial value immediately', () => {
+    const { result } = renderHook(() => useDebounce('hello', 500));
+
+    expect(result.current).toBe('hello');
+  });
+
+  it('updates value after specified delay', () => {
+    const { result, rerender } = renderHook(
+      ({ value, delay }) => useDebounce(value, delay),
+      { initialProps: { value: 'initial', delay: 300 } },
+    );
+
+    // Change the value
+    rerender({ value: 'updated', delay: 300 });
+
+    // Before the delay, still the old value
+    expect(result.current).toBe('initial');
+
+    // Advance time past the delay
+    act(() => {
+      vi.advanceTimersByTime(300);
+    });
+
+    expect(result.current).toBe('updated');
+  });
+
+  it('resets timer on rapid value changes', () => {
+    const { result, rerender } = renderHook(
+      ({ value }) => useDebounce(value, 300),
+      { initialProps: { value: 'a' } },
+    );
+
+    // Rapid changes
+    rerender({ value: 'b' });
+    act(() => { vi.advanceTimersByTime(100); });
+
+    rerender({ value: 'c' });
+    act(() => { vi.advanceTimersByTime(100); });
+
+    rerender({ value: 'd' });
+    act(() => { vi.advanceTimersByTime(100); });
+
+    // Not enough time has passed since last change -- still shows initial
+    expect(result.current).toBe('a');
+
+    // Advance past the delay from the last change
+    act(() => { vi.advanceTimersByTime(200); });
+
+    // Now it should show the final value
+    expect(result.current).toBe('d');
+  });
+
+  it('uses default delay of 300ms when not specified', () => {
+    const { result, rerender } = renderHook(
+      ({ value }) => useDebounce(value),
+      { initialProps: { value: 'start' } },
+    );
+
+    rerender({ value: 'end' });
+
+    act(() => { vi.advanceTimersByTime(299); });
+    expect(result.current).toBe('start');
+
+    act(() => { vi.advanceTimersByTime(1); });
+    expect(result.current).toBe('end');
+  });
+});
