+++ src/styles/tailwind.css
@@ -156,3 +156,13 @@
 .animate-page-in {
   animation: page-in 0.2s ease-out both;
 }
+
+/* Toast fade-in */
+@keyframes fade-in {
+  from { opacity: 0; transform: translateY(8px); }
+  to { opacity: 1; transform: translateY(0); }
+}
+
+.animate-fade-in {
+  animation: fade-in 0.2s ease-out both;
+}
